MOVEONLY: Move void UpdateTime() from pow.o to miner.o
